ActiScore Complete/AI Legal Document Summarizer & Query Assistant/requirements.txt
flask
transformers
torch
sentence-transformers
faiss-cpu
pillow
pytesseract
pdf2image
python-multipart
reportlab
numpy